feat: split tsconfig and eslint config for node and browser
