+++ .eslintrc.cjs
@@ -1,11 +1,5 @@
 module.exports = {
   root: true,
-  env: {
-    node: true,
-    browser: true,
-    jest: true,
-    webextensions: true,
-  },
   parser: '@typescript-eslint/parser',
   extends: [
     'plugin:@typescript-eslint/recommended',
