+++ __tests__/.eslintrc.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+    extends: "../.eslintrc.cjs",
+    env: {
+        node: true,
+        jest: true
+    },
+}
