+++ src/common/.eslintrc.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+    extends: "../../.eslintrc.cjs",
+    env: {
+        node: true,
+    },
+}
