+++ src/common/consts/network.ts
