+++ src/server/services/stateflow.service.ts
@@ -1,9 +1,9 @@
 import { StateflowServer } from '@ubi/ier-stateflow/dist/stateflow-server'
-import { STATEFLOW_PORT } from '@/server/consts/network'
+import { STATEFLOW_PORT } from '@/common/consts/network'
+import type { StateflowSpecification } from '@/common/types/stateflow'
 import { Logger } from '@/server/libs/logger'
 import type { Service } from '@/server/libs/service'
 import { getStateflowLoggerChannel } from '@/server/libs/stateflow-logger'
-import type { StateflowSpecification } from '@/server/types/stateflow'
 
 export class StateflowService implements Service {
   async main() {
