+++ src/tab-manager-extension/.eslintrc.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+    extends: "../../.eslintrc.cjs",
+    env: {
+        browser: true,
+        webextensions: true,
+    },
+}
